move entrypoints to right place
